+++ src/base/chat/legend.js
@@ -1,0 +1,41 @@
+wrap(() => {
+  const setting = settings.register({
+    name: 'Announcement',
+    key: 'underscript.announcement.legend',
+    options: ['Chat', 'Toast', 'Both', 'Hidden'],
+    default: 'Toast',
+    type: 'select',
+    page: 'Chat',
+    category: 'Legendary User',
+  });
+
+  const friends = settings.register({
+    name: 'Friends Only',
+    key: 'underscript.announcement.legend.friendsOnly',
+    page: 'Chat',
+    category: 'Legendary User',
+  });
+
+  eventManager.on('preChat:getMessageAuto', function legend({message: text}) {
+    const index = text.indexOf(' has just reached ');
+    if (this.canceled || !~index) return;
+    const handling = setting.value();
+    const friendsOnly = friends.value();
+    if (handling === 'Chat' && !friendsOnly) return; // All default
+    const hidden = handling === 'Hidden';
+    this.canceled = hidden || handling === 'Toast';
+    if (hidden) return;
+    const user = text.substring(0, index);
+    if (friendsOnly && !fn.isFriend(user)) {
+      this.canceled = true;
+      return;
+    }
+    fn.toast({
+      text,
+      css: {
+        color: 'yellow',
+        footer: {color: 'white'},
+      },
+    });
+  });
+});