+++ changelog.md
@@ -5,6 +5,8 @@
 1. Change log modifications
 1. No longer display patch notes when notifying about update
 1. @O now tags opponent
+1. Added options for legend rank announcement
+1. Added some settings
 
 ## Version 0.23.0-4
 ### New Features